MAINT: renames _beta._method to _beta._pipeline
