+++ q2_diversity/_beta/__init__.py
@@ -6,9 +6,9 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from ._method import (beta_phylogenetic, beta,
-                      all_phylo_metrics, all_nonphylo_metrics,
-                      all_metrics)
+from ._pipeline import (beta_phylogenetic, beta,
+                        all_phylo_metrics, all_nonphylo_metrics,
+                        all_metrics)
 from ._visualizer import (bioenv, beta_group_significance, mantel, adonis)
 from ._beta_rarefaction import beta_rarefaction
 from ._beta_correlation import beta_correlation
